# backend/apidocumentation.yaml
openapi: 3.0.0
info:
  title: Large Project API
  version: 1.0.0
  description: API documentation for Group 7's Large Project
servers:
  - url: http://leandrovivares.com
    description: Production server
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [First_Name, Last_Name, Username, Password, Email]
              properties:
                First_Name:
                  type: string
                Last_Name:
                  type: string
                Username:
                  type: string
                Password:
                  type: string
                  minLength: 5
                Email:
                  type: string
                  format: email
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or duplicate username/email
        '500':
          description: Server error
...
  /api/courses:
    get:
      summary: Get all courses
      tags: [Courses]
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
    post:
      summary: Add a course
      tags: [Courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Name, Code]
              properties:
                Name:
                  type: string
                Code:
                  type: string
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Name and Code are required
        '500':
          description: Server error

  # ── Professors ───────────────────────────────────────────────────────────────
  /api/professors:
    get:
      summary: Get all professors
      tags: [Professors]
      responses:
        '200':
          description: List of professors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Professor'
    post:
      summary: Add a professor
      tags: [Professors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [First_Name, Last_Name]
              properties:
                First_Name:
                  type: string
                Last_Name:
                  type: string
      responses:
        '201':
          description: Professor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Professor'
        '400':
          description: First name and last name are required
        '500':
          description: Server error
...
  /api/ratings:
    post:
      summary: Submit a rating
      tags: [Ratings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [User, Course, Option_A_Count, Option_B_Count, Option_C_Count, Option_D_Count, Option_E_Count]
              properties:
                User:
                  type: string
                Course:
                  type: string
                Professor:
                  type: string
                Option_A_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_B_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_C_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_D_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_E_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Professor_Rated:
                  type: boolean
                  default: false
                Option_F_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_G_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_H_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_I_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
                Option_J_Count:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '201':
          description: Rating created
        '400':
          description: Validation error or already rated
        '404':
          description: User not found
        '500':
          description: Server error
...
  /api/createQuestionnaire:
    post:
      summary: Create a questionnaire
      tags: [Questionnaires]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [User, Course, Question, Option_A, Option_B]
              properties:
                User:
                  type: string
                Course:
                  type: string
                Professor:
                  type: string
                Question:
                  type: string
                Option_A:
                  type: string
                Option_B:
                  type: string
                Option_C:
                  type: string
                Option_D:
                  type: string
      responses:
        '201':
          description: Questionnaire created
        '400':
          description: "write a question / add a MC option / User and Course are required"
        '500':
          description: Server error
...
components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        First_Name:
          type: string
        Last_Name:
          type: string
        Username:
          type: string
        Email:
          type: string
        Date_Created:
          type: string
          format: date-time
        Answered_Questionnaires:
          type: array
          items:
            type: string
        Rated_Courses:
          type: array
          items:
            type: string
        Rated_Courses_Professors:
          type: array
          items:
            type: object
            properties:
              Course:
                type: string
              Professor:
                type: string
    Course:
      type: object
      properties:
        _id:
          type: string
        Name:
          type: string
        Code:
          type: string
    Professor:
      type: object
      properties:
        _id:
          type: string
        First_Name:
          type: string
        Last_Name:
          type: string